requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
sqlparse==0.4.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.3
